refactor : auth 사용하는 로직 분리
